Setup game rules.

+++ FugoGames/Assets/Main/Scripts/Game/Deck.cs
@@ -1,10 +1,13 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Main.Scripts.Game
 {
     public class Deck
     {
+        public bool HasCards => CardsCount > 0;
         public int CardsCount => Cards.Count;
         
         public List<Card> Cards = new();
@@ -16,20 +19,27 @@ namespace Main.Scripts.Game
         
         public List<Card> Draw(int count)
         {
-            var drawnCards = Cards.Take(count).ToList();
+            var cardsCount = Math.Min(count, Cards.Count);
+            if (cardsCount <= 0)
+            {
+                return new List<Card>();
+            }
+            
+            var drawnCards = Cards.Take(cardsCount).ToList();
             Cards.RemoveRange(0, count);
+            
             return drawnCards;
         }
         
-        public void AddCard(List<Card> cards)
+        public void AddCards(List<Card> cards)
         {
             for (var i = 0; i < cards.Count; i++)
             {
-                AddCard(cards[i]);
+                AddCards(cards[i]);
             }
         }
         
-        public void AddCard(Card card)
+        public void AddCards(Card card)
         {
             Cards.Add(card);
         }
@@ -39,14 +49,10 @@ namespace Main.Scripts.Game
             Cards.Remove(selectedCard);
         }
         
+        [CanBeNull]
         public Card Last()
         {
-            if (CardsCount <= 0)
-            {
-                return null;
-            }
-            
-            return Cards[^1];
+            return !HasCards ? null : Cards[^1];
         }
         
     }
